+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+* Download permitted only if status processed
+* Remove error log
+
 ### 1.0.1: October 31st, 2018
 * Fix Github branch to allow updates